feat: :sparkles: Added navigation to photos by task page
