+++ src/components/jurado/TasksGallery.tsx
@@ -1,5 +1,6 @@
-import { actions } from "astro:actions";
+import { navigate } from "astro:transitions/client";
 import { useEffect, useState } from "react";
+import usePhotos from "../../hooks/usePhotos";
 import type { Photo, Task } from "../../types";
 import "./styles/TasksGallery.css";
 
@@ -18,34 +19,20 @@ const TasksGallery = ({
   refetchIntervalMs,
   colors,
 }: Props) => {
-  const [photos, setPhotos] = useState<Photo[]>(initialPhotos);
+  const { photos } = usePhotos({ initialPhotos, refetchIntervalMs });
   const [groupedTasks, setGroupedTasks] = useState<TaskWithPhotos[]>([]);
-
-  useEffect(() => {
-    const intervalId = setInterval(() => fetchPhotos(), refetchIntervalMs);
-    return () => clearInterval(intervalId);
-  }, []);
 
   useEffect(() => {
     const tasksGrouped = tasks.map((task) => {
       const taskPhotos = photos.filter((photo) => photo.taskId === task.id);
-      return { ...task, photos: taskPhotos };
+      return { ...task, photos: [...taskPhotos] };
     });
 
     setGroupedTasks(tasksGrouped);
-  }, [photos]);
+  }, [tasks, photos]);
 
-  const fetchPhotos = async () => {
-    try {
-      const data = await actions.getPhotos.orThrow();
-
-      setPhotos(data.photos);
-    } catch (error) {
-      // TODO: more UI friendly error handling
-      alert(
-        "Ha ocurrido un error al obtener las imágenes. Por favor, recarga la página.",
-      );
-    }
+  const handleTaskClick = (task: TaskWithPhotos) => {
+    navigate(`/jurado/tasks/${task.id}`);
   };
 
   return (
@@ -56,14 +43,17 @@ const TasksGallery = ({
           className="task-gallery-item"
           style={{ backgroundColor: colors[i % colors.length] }}
         >
-          <h3>
-            {task.text}
-            <span> {task.photos.length} fotos</span>
-          </h3>
+          <header onClick={() => handleTaskClick(task)}>
+            <h3>{task.text}</h3>·<span>{task.photos.length} fotos</span>
+            <p>Ver más...</p>
+          </header>
 
           <div className="task-photos">
-            {task.photos.map((photo, i) => (
-              <picture className={i === 0 ? "active" : ""}>
+            {task.photos.map((photo, index) => (
+              <picture
+                key={photo.public_id}
+                className={index === 0 ? "active" : ""}
+              >
                 <img src={photo.secure_url} />
               </picture>
             ))}
