+++ src/components/jurado/styles/TasksGallery.css
@@ -8,39 +8,61 @@
   margin-top: -2rem;
   padding: 1.5rem 2rem 3rem 2rem;
 
-  border: 2px solid var(--color-bg-dark);
-  border-top-left-radius: 2rem;
-  border-top-right-radius: 2rem;
+  border: 4px solid var(--color-bg-dark);
+  border-radius: 2rem;
   background: var(--color-bg);
 }
 
-.task-gallery-item h3 {
-  margin: 0 0 1rem 0;
+.task-gallery-item header {
+  margin-bottom: 1.5rem;
+
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+
+  color: var(--color-text-muted);
+}
+
+.task-gallery-item header h3 {
+  margin: 0;
 
   font-size: var(--text-size-md);
   font-weight: 600;
   color: var(--color-text);
 }
 
-.task-gallery-item h3 span {
-  float: right;
+.task-gallery-item header span {
+  margin: 0;
   font-size: var(--text-size-sm);
   font-weight: 400;
-  color: var(--color-text-muted);
+}
+
+.task-gallery-item header p {
+  margin: 0;
+  margin-left: auto;
+
+  font-size: var(--text-size-sm);
+  font-weight: 500;
+  cursor: pointer;
 }
 
 .task-photos {
   display: flex;
   align-items: end;
   gap: 1rem;
+  height: 150px;
+
+  overflow-x: hidden;
 }
 
 .task-photos picture {
   width: 100px;
   height: 100px;
 
-  border: 1px solid var(--color-grey);
+  flex: 0 0 auto;
+
   border-radius: 1rem;
+  box-shadow: var(--shadow-sm);
 
   overflow: hidden;
 }
